tdf#134610 sw: fix incorrect position of math object from DOCX
Regression from commit 9ace50192e450111638be063440acb6b3d1f552d
(tdf#130362 sw: fix anchoring of inline math objects, 2020-02-04), the
problem was that the SwFlyFrameAttrMgr ctor wanted to set the anchor
type to at-char, and then later
DocumentContentOperationsManager::InsertEmbObject() wanted to undo that
for math objects, but this did not play nicely with objects imported
from DOCX.

So don't set and clear the anchor type, rather set it conditionally in
the first place. This allows setting the anchor type in writerfilter/
before insertion, and then all of 1) docx import (depends) 2) insert of
Math objects (as-char) 3) insert of images (at-char) are working.

Change-Id: I94d82c12f30d069426db1bab70c456cadf1c91ef
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/99559
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins