Validate SELECT range
2d742fed1f6dd2298951fbb9d1cde96100d56a4a "Reliably set up controls of hyperlink
dialog in constructor" fixed an issue that could cause Python UITest code to
issue a bogus SELECT request that would fire an assert in
rtl_uString_newFromSubString.  Even if that issue is fixed now, it is probably a
good idea to validate the requested range here.

(An alternative would be to validate it in the underlying shell's SelectText,
but that function is also called from internal code (which presumably already
ensures that it is passing valid arguments), so it is probably better to stay
with that function's narrow interface.)

It would probably be nice if this function reported failure (by throwing a
css::lang::IllegalArgumentException, say) instead of silently (modulo SAL_WARN)
clamping the range, but it is called from Scheduler::ProcessTaskScheduling
(vcl/source/app/scheduler.cxx) in a try/catch block that prohibits all
exceptions.

Change-Id: I5b7b4255861766a81a81501e391b1ff4e09b7db6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/99933
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>