Avoid double init of sidebar controls
The welded sidebar contains only controls with officecfg based
registration. Such controls are created by the service manager,
which already calls the initialize method on its own, so there
is no point in doing it again. This code was copy-pasted from
the vcl based sidebar code path, but there it is guarded by a
condition to not apply to such controls since commit 62a951eb
("tdf#103624 Avoid double init of controls in SidebarToolBox").

Change-Id: I926f67f65b78799b6c3929184010be8af495817f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/100448
Tested-by: Jenkins
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>