tdf#123218 tdf#108067 config key ReverseXAxisOrientationDoughnutChart
for OOXML compatibility of newly created doughnut charts.

The X axis orientation of doughnut charts is the opposite of the primary
writing direction in LibreOffice but not in Microsoft Office. Default
value is "true" to keep current behavior. Using "false", the inner and
outer data series of new doughnut charts are the same, as after DOCX
export.

Change-Id: If431d5717e70599f07231bd673cd90c196450ae6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/100417
Tested-by: László Németh <nemeth@numbertext.org>
Reviewed-by: László Németh <nemeth@numbertext.org>