tdf#104596 sw layout: wrap in header only if bLayoutInCell
This patch depends on tdf#77794's
7.1 commit 07ac6e2b6134f44e6bdec13d192eaa948e62e940.

For more details about the issue in general,
see the verbose comments in this bug's previous
7.1 commit 6a5165bd729c9bd959048c18936b8c7e703c3486.

This patch is related to CompatibilityMode < 15.
Unfortunately, the previous patch didn't work
with older Word 2010 versions of the file,
which _shouldn't_ wrap non-LayoutInCell table-anchored flies.

Unfortunately, now that different behaviour is necessary
for different Word compat levels,
it no longer allows a nice way for Writer to handle
this natively. So since it would be very unlikely for a
user to create a document like this (since the necessary
"keep inside text boundaries" is off by default in Writer,
but is forced on by definition in Word 2013+),
I'm removing the compatibility flag I added in 7.1,
and its related unit test.

[To do this natively would probably require enabling
the IsFollowingTextFlow property by default in SW.
That sounds very dangerous since this property
is not restricted to IsInTable layout situations.
This property has been around since at least LO 3.5.]

Change-Id: I70da016cb68f515924ed6c17085bf73a9e1c5492
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/100684
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>