solenv: fix typo in CppunitTest
Added in commit ee67bedaf424734ab0f2cb6c2061f573488d537d (Default all
tests to run with the svp plugin, 2018-11-23), if svp is to be avoided,
it has to be filtered out from the environment variables, not filtered
for.

With this, the unexpected SAL_USE_VCLPLUGIN=gen is gone from the
generated cmdline when using gb_CppunitTest_use_vcl_non_headless.

Change-Id: Ib666f3df007898165f2019f0a9b0677f679aa6e1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/100742
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>