tdf#135457 sw: fix invalidation of SwXTextField
This was supposed to happen from SwNodes::ChgNode() when moving to the
undo nodes array.

There are fields with a range that contains other fields, when the outer
field's range is deleted, the contained field moves to the undo array
and its position is no longer valid to be given out by
SwXTextField::getAnchor().

(regression from 17426ecf474971306cacd652a3fe1d7b646237f8)

Change-Id: I38c8b3393a777ca8dc8739b5ca4233a7f124961e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/100749
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>