tdf#135366 ODF import of line and fill of data labels
LibreOffice has line and fill properties of data labels in a chart in
properties of kind "LabelBorderWidth" or "LabelFillColor" at a data
point or, as defaults for the points, at a series.

But ODF has such information in a <style:style> element, which is
refered by a <chart:data-label> element. That one can be child of a
<chart:data-point> and child of a <chart:series> element.

Microsoft Office correctly uses the <chart:data-point> element and its
style for the line and fill properties of data labels. Up to now LO
cannot import such information and does not write the ODF elements.
Instead LibreOffice reads and writes attributes in 'loext' namespace.

Using the "LabelFoo" properties was implemented by Kohei Yoshida,
July 2014. Although there is no published service, these properties
can be used in macros. Because they are now available since six
years, the decision was to keep this internal model and convert on
import and export.

This patch implements the import of the ODF fill and line properties
from the <chart:data-label> element and converts them to the internal
used properties.

LibreOffice has currently only implemented a few of the possible line
and fill properties. When more are implemented, their <ODF,LabelFoo>
pairs need to be added to the array aApiToLabelFooPairs, further
adaptions are not needed.

The <chart:data-label> contains in addition the absolute position of
a data label. LibreOffice has internally only a position offset
relative to the regular position of the label. The conversion of the
position is not included in the patch.

Change-Id: I5fd868945585971dac3dd945804a4a2951b8c12d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/101194
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>