Access2Base - tdf#136063 Workaround Basic missing argument handling
When an argument is missing in a call to a Basic function
the IsMissing() function applied on the argument
should respond True.
However sometimes(?) the argument contains instead "Error 448"
making IsMissing responding False.

This commit puts an explicit argument in the function call
avoiding the different behaviour of the Basic interpreter
in LO 7.0 vs. previous releases.

Change-Id: Ic90fb7ec9b6e520cc762a441ca383b54c5d4db16