related tdf#132149 ww8 export: unit test to prevent bad fix
The unit test I am using for this patch writes out landscape
attribute, but provides portrait width/height values.

I was tempted to just "fix" the values on export, but this
existing document shows that w/h trump p/l in LO
(and Word does the same). So that should also round-trip,
and this test will ensure that keeps happening.

Change-Id: Ib55cb799462abd1039ce7c1c935b3f66761a5dc2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/101479
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>