Fix ooo38104-1.sxw crash after 01da9cfbfceed33d05e77e4a759ffa8b7b52d7b4
Unlike the c123bf commit, this commit does not cause the crash that
was caught by the crash-testing system.

(The crash could be reproduced by:
    wget -O ooo38104-1.sxw https://bz.apache.org/ooo/attachment.cgi?id=19889
    ./instdir/program/soffice.bin --headless --convert-to docx ./ooo38104-1.sxw
)

In this commit, I reinstate the "early return" in
SwView::ReadUserDataSequence() that I dropped in the c123bf commit,
but instead move the SelectShell() call earlier, so that it will be
executed before the potential early return.

The problem that we try to fix here is the one that the fresh
CppunitTest_sw_updateall_object_replacements checks, so to reproduce
that problem, revert both this commit and c123bf, and then run that
unit test.

Change-Id: I6c728b75a2f172b75fbf2ad00c019c32aecc19f8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/101531
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>