Fix crash when entering newline in starmath experimental editor
Currently things crash when adding a newline using the experimental
editor in Math Formula editor. It looks like this is happening because
the newly inserted line was getting destroyed when the unique_ptr goes
out of scope.

This change releases the unique_ptr so it doesn't get destroyed. As it
is now part of the tree it should be destroyed elsewhere.

Change-Id: Icfe2c0289c12302d39883dc04cef0351467aa845
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/101737
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>