UITest_writer_tests: don't assert exact page count at test_tdf135018() end
For example

cd sw; make -sr UITest_writer_tests UITEST_TEST_NAME="trackedChanges.trackedchanges.test_tdf135018"

fails for me like this:

    self.assertEqual(18, document.CurrentController.PageCount)
AssertionError: 18 != 17

This is likely because the test document is not carefully preparted to
only use bundled fonts. The test mentions that the primary point is to
make sure we don't crash, so just give up on asserting the exact page
number.

Change-Id: I212c02995fd487526d4e2fab7b351d383f9b8459
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/101980
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>