avoid temporary SkImage when merging bitmaps in Skia (tdf#136244)
The original idea was to create a temporary SkImage in order to first
merge the bitmap and its alpha bitmap, otherwise scaling already
while merging them could introduce artefacts because of smoothscaling
the alpha. But SkShader use blends the bitmap and alpha values
before the scaling, so this is actually not necessary.

Change-Id: I4e351611e3c33530dd5326c542b0a191955b5109
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/102068
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>