do not queue up too many Skia image drawing operations (tdf#136369)
The document tries to draw a larger number of tiny bitmaps,
which results in Skia trying to allocate many tiny GPU images,
which slows it down (something quadratic there?) and it may
eventually run out of memory. So force a flush if there are
too many such operations unflushed.
This is somewhat similar to what was done in the past by
891cee4a17c0239efc2b4c9b38a62510d6c10706 .

Change-Id: Ib0309ea9c7ac6c80d69ceed2922e885a2aa7f2d8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/102166
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>