tdf#135149 sw: fix deleting textbox of as-char shapes
Regression from 815e3c3b7207aa0970c0d93eecde342b14494dcf (Add
SwTextBoxHelper::findShapes, 2014-06-24), the problem was that in case
the shape is as-char anchored, then DeleteSelection() returns false, and
so the textboxes of the deleted shapes are not deleted.

There is no real reason why deletion should be conditional on the
DeleteSelection() result, so just move it outside that block, so
textboxes are always deleted.

Change-Id: Ib3bd444e9e96f9de0f68a9162ab39255ec5a6807
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/102198
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>